Moving_client_project

+++ API/Controllers/StoreController.cs
@@ -3,8 +3,10 @@ using AutoMapper;
 using Core.DTOs;
 using Core.Entities;
 using Core.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace API.Controllers
 {
@@ -22,27 +24,35 @@ namespace API.Controllers
 
         [HttpPost]
         [Route("Create")]
+        //[Authorize(Roles ="Vendor")]
         public async Task<IActionResult> CreateStore(StoreDto storeDto)
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            Store store = new Store
-            {
-                Name = storeDto.Name,
-                Description = storeDto.Description,
-                CreatedDate = DateTime.Now,
-                ModifiedDate = DateTime.Now,
-                IsActive = false,
-                //Address = storeDto.Address,
-                PhoneNumber = storeDto.PhoneNumber,
-                EmailAddress = storeDto.EmailAddress
-            };
+            //var userId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            string? userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userId != storeDto.VendorId)
+                return Forbid("Something wrong with user authorization!");
 
-            bool result = await _context.Add(store);
+            Store store = _mapper.Map<Store>(storeDto);
+            //    new Store
+            //{
+            //    Name = storeDto.Name,
+            //    Description = storeDto.Description,
+            //    CreatedDate = DateTime.Now,
+            //    ModifiedDate = DateTime.Now,
+            //    IsActive = false,
+            //    //Address = storeDto.Address,
+            //    PhoneNumber = storeDto.PhoneNumber,
+            //    EmailAddress = storeDto.EmailAddress,
+            //    VendorId = storeDto.VendorId
+            //};
+
+            Store result = await _context.Add(store);
 
 
-            if (result)
+            if (result != null)
                 return Ok("Store Created Successfully!");
             else
                 return BadRequest("Something wend wrong!");
@@ -52,7 +62,8 @@ namespace API.Controllers
         [Route("Stores")]
         public async Task<ActionResult<IEnumerable<Store>>> GetAll()
         {
-            return Ok(await this._context.GetAll());
+            List<StoreDto> stores = _mapper.Map<List<StoreDto>>(await this._context.GetAll());
+            return Ok(stores);
         }
 
         [HttpGet]
@@ -74,10 +85,17 @@ namespace API.Controllers
         [Route("Update/{id}")]
         public async Task<ActionResult<Store>> UpdateStore(Guid id, [FromBody] StoreDto storeDto)
         {
+            
             Store existingStore = await this._context.GetByIdAsync(id);
 
+           
             if (existingStore == null)
                 return BadRequest("Store doesn't exist!");
+
+            string? vendorId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (existingStore.VendorId != vendorId)
+                return Forbid("Unautorized vendor!");
 
             UpdateEntitiesHelper.UpdateStore(existingStore, storeDto);
 
@@ -96,6 +114,11 @@ namespace API.Controllers
             Store store = await this._context.GetByIdAsync(id);
             if (store != null)
                 return BadRequest("Store doesn't exist!");
+
+            string? vendorId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (store?.VendorId != vendorId)
+                return Forbid("Unautorized vendor!");
 
             await this._context.Delete(store);
